Aura stacks now have float support in advskills

+++ Skua.Core.Models/Skills/AdvancedSkillJsonModels.cs
@@ -31,7 +31,7 @@ public class SkillRuleJson
 
     [JsonPropertyName("value")]
     [JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingNull)]
-    public int? Value { get; set; }
+    public float? Value { get; set; }
 
     [JsonPropertyName("comparison")]
     [JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingNull)]
